refactor: renamed blog folder
